+++ src/resources/Channelzones.ts
@@ -2,23 +2,35 @@ import TelnyxResource from '../TelnyxResource';
 const telnyxMethod = TelnyxResource.method;
 
 export const Channelzones = TelnyxResource.extend({
-  path: 'channelzones',
-  includeBasic: ['list', 'retrieve', 'delete'],
+  path: 'channel_zones',
+  includeBasic: ['list'],
 
-  GetChannelZone: telnyxMethod({
-    method: 'GET',
-    path: '/channel_zones/{channel/zone/id}',
+  update: telnyxMethod({
+    method: 'PATCH',
+    path: '/{channel_zone_id}',
+    urlParams: ['channel_zone_id'],
   }),
-  UnassignPhoneNumber: telnyxMethod({
+
+  retrieve: telnyxMethod({
+    method: 'GET',
+    path: '/{channel_zone_id}',
+    urlParams: ['channel_zone_id'],
+  }),
+
+  createPhoneNumber: telnyxMethod({
+    method: 'POST',
+    path: '/{channel_zone_id}/channel_zone_phone_numbers',
+    urlParams: ['channel_zone_id'],
+  }),
+
+  listPhoneNumbers: telnyxMethod({
+    method: 'GET',
+    path: '/{channel_zone_id}/channel_zone_phone_numbers',
+  }),
+
+  delPhoneNumber: telnyxMethod({
     method: 'DELETE',
-    path: '/channel_zones/{channel/zone/id}/channel_zone_phone_numbers/{phone_number}',
-  }),
-  GetChannelZones: telnyxMethod({
-    method: 'GET',
-    path: '/channel/zones',
-  }),
-  GetPhoneNumbers: telnyxMethod({
-    method: 'GET',
-    path: '/channel_zones/{channel_zone_id}/channel_zone_phone_numbers',
+    path: '/{channel_zone_id}/channel_zone_phone_numbers/{phone_number}',
+    urlParams: ['channel_zone_id', 'phone_number'],
   }),
 });